# testsuite: remove sched-dummy, migrate testing to sched-simple