… t/t2203-job-manager-single.t
… t/t2204-job-manager-dummysched-unlimited.t
… t/t2204-job-manager-unlimited.t
… t/t2205-job-manager-annotate.t
… t/t2215-job-manager-priority-order-unlimited.t