per discussion in #3455 

- some misc fixups, including a pretty big bug i found in some tests
- made `single` and `unlimited` mode tests more consistent to each other in the testsuite.  Unsure of legacy reasons why tests were different (e.g. in `single` mode `resource_summary` annotation was not output in `sched-dummy`).
- add a LOG_DEBUG message to `hello_cb` in `sched-simple`.  Could be made to output only on debugging circumstances, but thought it was ok to leave in for all circumstances.
- add a DEBUG_FAIL_ALLOC test hook to `sched-simple` to automatically fail the next allocation
- add a DEBUG_ANNOTATE_REASON_PENDING hook to inform `sched-simple` to annotate with `reason_pending` and `jobs_ahead` annotations
- update all tests to use `sched-simple` over `sched-dummy`
- remove `sched-dummy` and rename some files