@grondo 2021-01-08
Before I delve in (later), noticed inconsistent use of capitalization in commit messages, e.g. `￼sched-simple: Fix code tabbing`, vs `￼sched-simple: fix error message typo`. (I'm sure I'm guilty of this as well). I prefer the 2nd (no capitalization), since commit subject is known not to be a full sentence anyway.