@chu11 2021-01-08
rebased and re-pushed.  I had a number of builder failures with loading/unloading modules, which I couldn't figure out.  Possibly something racy I was missing.

Then it occurred to me, with the removal `sched-dummy`, I no longer have to manually load/unload `job-ingest`, `job-info`, `job-manager`, I can run under the `job` test personality instead.  So I adjusted.  We'll see if builders work this time.