@chu11 2021-01-09
one builder failed w/ #3463 and another failed with a seemingly random collection of `flux-mini` failures.  I'd be worried it was something I did if it weren't for #3463 coming up at same time today.  In the `centos8 - py3.7` builder

```
Run src/test/checks-annotate.sh
Warning: Found 3 errors from 259 tests in testsuite
Error: not ok 8 - flux-mini batch: submit a series of jobs
Error: not ok 9 - flux-mini batch: job results are expected
Error: ERROR: t2701-mini-batch.t - exited with status 1
./t/t2701-mini-batch.output
./t/t2701-mini-batch.log
Error: not ok 30 - flux-shell: shell can launch flux
Error: ERROR: t2601-job-shell-standalone.t - exited with status 1
./t/t2601-job-shell-standalone.log
Error: not ok 7 - flux-mini alloc runs one broker per node by default
Error: ERROR: t2702-mini-alloc.t - exited with status 1
./t/t2702-mini-alloc.output
./t/t2702-mini-alloc.log
```

```
  expecting success: 
  	id1=$(flux mini batch --flags=waitable -n1 batch-script.sh) &&
  	id2=$(flux mini batch --flags=waitable -n4 batch-script.sh) &&
  	id3=$(flux mini batch --flags=waitable -N2 -n4 batch-script.sh) &&
  	run_timeout 60 flux job wait --verbose --all
  
  f2k6U2Yo: job completed successfully
  f2cqRYWs: job completed successfully
  not ok 8 - flux-mini batch: submit a series of jobs
  
  expecting success: 
  	test_debug "grep . flux-*.out" &&
  	grep "size=1 nodes=1" flux-${id1}.out &&
  	grep "size=1 nodes=1" flux-${id2}.out &&
  	grep "size=2 nodes=2" flux-${id3}.out
  
  size=1 nodes=1
  size=1 nodes=1
  not ok 9 - flux-mini batch: job results are expected
```

(no .output file?)
```
  FAIL: t2601-job-shell-standalone.t 30 - flux-shell: shell can launch flux
  #	
  #		flux jobspec srun -N1 -n8 flux start flux getattr size >j8flux &&
  #		${FLUX_SHELL} -v -s -r 0 -j j8flux -R R8 39 \
  #			>flux.out 2>flux.err &&
  #		grep -x "0: 8" flux.out
  #	
```

```
expecting success: 
  	$runpty -o multi.out flux mini alloc -n5 flux lsattr -v &&
  	test_debug "cat multi.out" &&
  	grep "size  *2" multi.out
  
  runpty: INFO: child exited with status 0x100
  not ok 7 - flux-mini alloc runs one broker per node by default`

```