@grondo 2021-01-12
I'm going to try restarting tests. It seemed like maybe Github runners were behaving strangely late last week.

Then it seems like it would be better to get this cleanup in sooner rather than later, so I'll take an immediate look.