@grondo 2021-01-12
Hm, there seems to be something wrong with the scheduler on this branch. I have a little job throughput test which reproducibly gets ~70 job/s on current master, but on this branch I'm seeing about 6 jobs/s.

It could be something in my environment, though I'm running both tests from different build directories on fluke.

However, @chu11, maybe you can try this out and see if the problem reproduces for you?

```python
#!/usr/bin/env python3
#
#
import sys
import time
import argparse

import flux
from flux import job
from flux.job import JobspecV1, JobID

try:
    from flux.progress import Bottombar
except:
    pass

parser = argparse.ArgumentParser(description="Run job throughput test")
parser.add_argument(
    "-n",
    "--njobs",
    type=int,
    metavar="N",
    help="Set the total number of jobs to run",
    default=100,
)
parser.add_argument(
    "-x",
    "--exec",
    help="Do not simulate execution, actually run jobs",
    action="store_true",
)
parser.add_argument(
    "-s",
    "--status",
    help="Add a status bar to bottom of terminal",
    action="store_true",
)
parser.add_argument(
    "-v",
    "--verbose",
    help="Log job events",
    action="store_true",
)
parser.add_argument("command", nargs=argparse.REMAINDER, default=["true"])

args = parser.parse_args()


class BulkRun:
    def __init__(self, handle, total, jobspec):
        self.handle = handle
        self.total = total
        self.jobspec = jobspec
        self.jobs = {}
        self.submitted = 0
        self.running = 0
        self.complete = 0
        self.bb = None

    def statusline(self, bb, width):
        return (
            f"{self.total:>6} jobs: {self.submitted:>6} submitted, "
            f"{self.running:>6} running, {self.complete} completed"
        )

    def event_cb(self, future, args, jobid):
        event = future.get_event()
        if event is not None:
            if args.verbose:
                print(f"{jobid}: {event.name}")
            if event.name == "submit":
                self.submitted += 1
                self.jobs[jobid][event.name] = event
            elif event.name == "start":
                self.running += 1
            elif event.name == "finish":
                self.running -= 1
                self.complete += 1
            elif event.name == "clean":
                self.jobs[jobid][event.name] = event
            if self.bb:
                self.bb.update()

    def submit_cb(self, future, args):
        try:
            jobid = JobID(future.get_id())
            self.jobs[jobid] = {}
            fut = job.event_watch_async(self.handle, jobid)
            fut.then(self.event_cb, args, jobid)
        except Exception as exc:
            print(f"Submission failed: {exc}", file=sys.stderr)

    def run(self, args):
        if args.status:
            self.bb = Bottombar(self.statusline).start()
        spec = self.jobspec.dumps()
        for i in range(0, args.njobs):
            job.submit_async(self.handle, spec).then(self.submit_cb, args)
        self.handle.reactor_run()

        if self.bb:
            self.bb.stop()

        return self


t0 = time.time()

#  Create a test jobspec
if not args.command:
    args.command = ["true"]
jobspec = JobspecV1.from_command(args.command)
if not args.exec:
    jobspec.setattr("system.exec.test.run_duration", "0.001s")

bulk = BulkRun(flux.Flux(), args.njobs, jobspec).run(args)

jobs = bulk.jobs

#  Get the job with the earliest 'submit' event:
first = jobs[min(jobs.keys(), key=lambda x: jobs[x]["submit"].timestamp)]

#  Get the job with the latest 'clean' event:
last = jobs[max(jobs.keys(), key=lambda x: jobs[x]["clean"].timestamp)]

script_runtime = time.time() - t0
job_runtime = last["clean"].timestamp - first["submit"].timestamp
jps = args.njobs / job_runtime
jpsb = args.njobs / script_runtime

print(f"number of jobs: {args.njobs}")
print(f"script runtime: {script_runtime:<6.3f}s")
print(f"job runtime:    {job_runtime:<6.3f}s")
print(f"throughput:     {jps:<.1f} job/s (script: {jpsb:5.1f} job/s)")
```

On master:
```
$ src/cmd/flux start sh -c 'flux module reload sched-simple unlimited; flux python ../flux-core.git/throughput.py -n 1024'
number of jobs: 1024
script runtime: 14.855s
job runtime:    14.706s
throughput:     69.6 job/s (script:  68.9 job/s)
```

On this branch:
```
$  src/cmd/flux start sh -c 'flux module reload sched-simple unlimited; flux python ../flux-core.git/throughput.py -n 1024'
number of jobs: 1024
script runtime: 184.284s
job runtime:    154.650s
throughput:     6.6 job/s (script:   5.6 job/s)
```

It seems to get much worse the more jobs are enqueued, so something in this PR is not scaling at all with the number of submitted jobs in the "unlimited" case.